+++ src/main/java/com/example/covid19api/service/DataService.java
@@ -176,56 +176,3 @@ public class DataService {
     }
 }
 
-//
-//    public List<LocationConfirmedData> confirmedDataByLocation(String file) {
-//        List<LocationConfirmedData> locationConfirmedDataList = new ArrayList<>();
-//        List<String[]> data = ReadCSV.readCSVFile(file);
-//        for (int i = 1; i < data.size(); i++) {
-//            String[] row = data.get(i);
-//            if (!row[0].equals("Recovered")) {
-//                int confirmed = Integer.parseInt(row[row.length - 1]);
-//                LocationConfirmedData locationConfirmedData = new LocationConfirmedData(row[1],
-//                                                                                        Optional.of(row[0]),
-//                                                                                        row[2],
-//                                                                                        row[3],
-//                                                                                        confirmed);
-//                locationConfirmedDataList.add(locationConfirmedData);
-//            }
-//        }
-//        return locationConfirmedDataList;
-//    }
-//
-//    public List<LocationDeathData> deathDataByLocation(String file) {
-//        List<LocationDeathData> locationDeathDataList = new ArrayList<>();
-//        List<String[]> data = ReadCSV.readCSVFile(file);
-//        for (int i = 1; i < data.size(); i++) {
-//            String[] row = data.get(i);
-//            if (!row[0].equals("Recovered")) {
-//                int deaths = Integer.parseInt(row[row.length - 1]);
-//                LocationDeathData locationDeathData = new LocationDeathData(row[1],
-//                                                                            Optional.of(row[0]),
-//                                                                            row[2],
-//                                                                            row[3],
-//                                                                            deaths);
-//                locationDeathDataList.add(locationDeathData);
-//            }
-//        }
-//        return locationDeathDataList;
-//    }
-//
-//    public List<LocationRecoveredData> recoveredDataByLocation(String file) {
-//        List<LocationRecoveredData> locationRecoveredDataList = new ArrayList<>();
-//        List<String[]> data = ReadCSV.readCSVFile(file);
-//        for (int i = 1; i < data.size(); i++) {
-//            String[] row = data.get(i);
-//            int recovered = Integer.parseInt(row[row.length - 1]);
-//            LocationRecoveredData locationRecoveredData = new LocationRecoveredData(row[1],
-//                                                                                    Optional.of(row[0]),
-//                                                                                    row[2],
-//                                                                                    row[3],
-//                                                                                    recovered);
-//            locationRecoveredDataList.add(locationRecoveredData);
-//        }
-//        return locationRecoveredDataList;
-//    }
-//}
